+++ examples/screenshot/src/main.rs
@@ -1,6 +1,7 @@
-// #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
 
 use eframe::egui::{self, ColorImage};
+use image;
 
 fn main() -> Result<(), eframe::Error> {
     let options = eframe::NativeOptions {
@@ -19,6 +20,7 @@ struct MyApp {
     continuously_take_screenshots: bool,
     texture: Option<egui::TextureHandle>,
     screenshot: Option<ColorImage>,
+    save_to_file: bool,
 }
 
 impl eframe::App for MyApp {
@@ -38,6 +40,12 @@ impl eframe::App for MyApp {
                     "continuously take screenshots",
                 );
 
+                if self.continuously_take_screenshots {
+                    if ui.button("save to 'top_left.png'").clicked() {
+                        self.save_to_file = true;
+                    }
+                }
+
                 ui.with_layout(egui::Layout::top_down(egui::Align::RIGHT), |ui| {
                     if self.continuously_take_screenshots {
                         if ui
@@ -48,9 +56,9 @@ impl eframe::App for MyApp {
                         } else {
                             ctx.set_visuals(egui::Visuals::light());
                         };
-                        frame.request_pixels();
+                        frame.request_screenshot();
                     } else if ui.button("take screenshot!").clicked() {
-                        frame.request_pixels();
+                        frame.request_screenshot();
                     }
                 });
             });
@@ -65,10 +73,24 @@ impl eframe::App for MyApp {
         });
     }
 
-    #[allow(unsafe_code)]
-    fn post_rendering(&mut self, _screen_size_px: [u32; 2], frame: &eframe::Frame) {
-        if let Some(pixels) = frame.frame_pixels() {
-            self.screenshot = Some(pixels)
+    fn post_rendering(&mut self, _window_size: [u32; 2], frame: &eframe::Frame) {
+        if let Some(screenshot) = frame.screenshot() {
+            if self.save_to_file {
+                let pixels_per_point = frame.info().native_pixels_per_point;
+                let region =
+                    egui::Rect::from_two_pos(egui::Pos2::ZERO, egui::Pos2 { x: 100., y: 100. });
+                let top_left_corner = screenshot.region(&region, pixels_per_point);
+                image::save_buffer(
+                    "top_left.png",
+                    top_left_corner.as_raw(),
+                    top_left_corner.width() as u32,
+                    top_left_corner.height() as u32,
+                    image::ColorType::Rgba8,
+                )
+                .unwrap();
+                self.save_to_file = false;
+            }
+            self.screenshot = Some(screenshot)
         }
     }
 }